requirements.txt
CoinMarketCapAPI==0.5
cryptography==42.0.7
cv-bridge==3.2.1
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
pcl-msgs==1.0.0
pendulum-msgs==0.20.3
pillow==10.3.0
platformdirs==4.2.1
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.22
pyparsing==3.1.2
pysqlite3==0.5.2
python-dateutil==2.9.0.post0
python-qt-binding==1.1.1
pytz==2024.1
stable-baselines3==2.3.2
